updated query to load deal_stage records
NOTE: In report 8 (RPT-8) deal_stage is joined directly to the deal table so this should be fine to join to for transaction_dim. Most of the references to deal_stage join from deal_summary. I don’t see any where they are joining through proposal as Griff discussed in the meeting this morning.